webapp/requirements.txt
Flask==2.0.1
torch==1.9.0
matplotlib==3.4.3
numpy==1.21.1
Pillow==8.2.0
pandas==1.3.1
tqdm==4.61.2
torchvision==0.10.0
